+++ miracle/analysis/tests/test_example.py
@@ -6,7 +6,7 @@ from miracle.analysis import example
 
 
 def test_main(db):
-    assert example.main(db)
+    assert example.main(db) is None
 
 
 def test_example(db):