+++ miracle/scripts/tests/test_analysis.py
@@ -2,5 +2,5 @@ from miracle.scripts import analysis
 
 
 def test_main(db):
-    assert not analysis.main(['dummy', '--script=foo'], _db=db)
-    assert analysis.main(['dummy', '--script=example'], _db=db)
+    assert analysis.main(['dummy', '--script=foo'], _db=db) == 1
+    assert analysis.main(['dummy', '--script=example'], _db=db) is None